+++ persist-restaurant-data/src/main/java/com/services/persistrestaurantdata/restaurant/repository/RestaurantRepository.java
@@ -1,8 +1,9 @@
-package com.services.persistrestaurantdata.restaurant;
+package com.services.persistrestaurantdata.restaurant.repository;
 
+import com.services.persistrestaurantdata.restaurant.model.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-@Repository
+@Repository("restaurantRepo")
 public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
 }
