+++ restaurant-data-api-client/src/main/java/com/services/fetchrestaurantdata/controller/RestaurantApiController.java
@@ -1,5 +1,7 @@
-package com.services.persistrestaurantdata.restaurant;
+package com.services.fetchrestaurantdata.controller;
 
+import com.services.fetchrestaurantdata.dto.RestaurantClientDto;
+import com.services.fetchrestaurantdata.service.RestaurantDataService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
@@ -12,11 +14,11 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/v1/restaurants")
+@RequestMapping("/api/v1/restaurants-data")
 @RequiredArgsConstructor
-public class RestaurantController {
+public class RestaurantApiController {
 
-    private final RestaurantRepository restaurantRepository;
+    private final RestaurantDataService dataService;
 
     @GetMapping("/all/restaurants")
     @Operation(summary = "fetch all restaurants from db", description = "fetch restaurants")
@@ -24,7 +26,7 @@ public class RestaurantController {
             @ApiResponse(responseCode = "200", description = "Successfully fetched restaurant data"),
             @ApiResponse(responseCode = "500", description = "Internal server error occurred")
     })
-    ResponseEntity<List<Restaurant>> findAll() {
-        return ResponseEntity.ok(restaurantRepository.findAll());
+    ResponseEntity<List<RestaurantClientDto>> findAll() {
+        return ResponseEntity.ok(dataService.getAllRestaurants());
     }
 }
