+++ restaurant-data-api-client/src/main/java/com/services/fetchrestaurantdata/dto/RestaurantClientDto.java
@@ -1,0 +1,17 @@
+package com.services.fetchrestaurantdata.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@Builder
+@AllArgsConstructor
+public class RestaurantClientDto {
+    Long id;
+    String name;
+    double rating;
+    String working_time;
+    String phone_number;
+    String address;
+}
