+++ restaurant-data-api-client/src/main/java/com/services/fetchrestaurantdata/service/RestaurantDataService.java
@@ -1,0 +1,10 @@
+package com.services.fetchrestaurantdata.service;
+
+import com.services.fetchrestaurantdata.dto.RestaurantClientDto;
+
+import java.util.List;
+
+public interface RestaurantDataService {
+
+    List<RestaurantClientDto> getAllRestaurants();
+}
