+++ restaurant-data-api-client/src/main/resources/application.yml
@@ -1,0 +1,2 @@
+server:
+  port: 8971
